requirements.txt
# application
flask
tweepy
python-dotenv
pymongo

# tests
pytest
pytest-watch
requests
vcrpy
pytest-vcr

# operations
gunicorn
